Refactor: Move amici.de_model to amici.de_model_components
... to make space to move amici.de_export.DEModel to amici.de_model
