+++ .github/workflows/test_sbml_semantic_test_suite.yml
@@ -9,7 +9,7 @@ on:
     paths:
       - .github/workflows/test_sbml_semantic_test_suite.yml
       - python/sdist/amici/de_export.py
-      - python/sdist/amici/de_model.py
+      - python/sdist/amici/de_model_components.py
       - python/sdist/amici/sbml_import.py
       - python/sdist/amici/import_utils.py
       - scripts/run-SBMLTestsuite.sh
